docs[prep]: MMA Judging Bot row Update
Feasibility: I put feasibility as 60, on the lower end. This is because judging a fight means judging at least 3 separate 5 minute rounds, which if a fight played at 60 fps were to be fed into a hypothetical algorithm, that would mean 54000 frames to analyze. I did not drop the score too low because the project is very scalable, as we could go from simply recognizing and counting certain aspects of MMA (such as striking) before fully integrating grappling (control time, positions, submission attempts). Getting the footage for fights "legally" will also be a problem depending on the fight.

Real-world impact: MMA is a mainstream but on the niche side sport. If an MMA judging bot could improve the sport, that would be useful, but at the end of the day sports are just entertainment, so I only rated it 50 (probably high)

Productive in Industry: Judging is basically notoriously bad and inconsistent for basically all combat sports. An ML algorithm wouldn't need to be incredibly reliable to outperform human judges and it would great because commissions will no longer have to pay geriatrics a wage to rob fighters, it would help MMA organizations avoid controversies, and a lot of fighters would not be screwed out of their win bonuses. I rated this 80, not 100 because at the end of the day there's still possibility for things to go wrong if the algorithm isn't recognizing what's happening properly and/or doesn't rate actions the same way the scoring criteria does.

Productive Personally: I think generally quite a lot of the group members enjoy watching MMA and think that making a software solution that would truly improve the sport would be great. Because of the uniqueness of the project and how easily understandable what it does, I gave this an 80, but not 100 because some members aren't quite on board with how technically feasible it is.

Economic Impact: Yes. The bot could be licensed to MMA organizations for a profit on the developers and it would save MMA organizations the costs of paying human judges. Since technically some people would be losing their jobs, I only gave this an 80.

Socially Responsible: I think this project doesn't really have any harms unless the algorithm actively doesn't function and it gets used anyways. It would also cause some judges to lose their jobs, so I gave this an 80.

Interesting Computationally: I think the mechanics of the algorithm (recognizing strikes, control, submission attempts) are very interesting but I deducted 30 marks because of how much data would have to be worked with. Blending the weighting on the different actions the fighter does in order to score a round would also be potentially very interesting. There's a very good chance the computation and data in this project ends up being both very challenging and tedious because we are applying an ML algorithm to video and not just still images.

Impressive Computationally: I gave this 100 because I think it's instantly recognizable if you have even passing knowledge of the sport how impressive and useful this project could be.

Interesting Data: I gave this an 80 tentatively. I need to do more research on how ML algorithms work on videos, but I can imagine it isn't terribly different from the many faces assignment/project from 4ML3 where we convert the images to grayscale and then work with them. I once again deducted a bunch of points because of how much data the project would require means it is likely to end up being quite tedious.

Impressive Data: I think generally explaining that the algorithm parses through every frame of a recording of a fight is impressive in a self-explanatory manner so I gave this 100.

Interesting System Integration: As discussed in the meeting, this project is easily scalable by parts - we could start by just recognizing strikes, then separate the strikes by whether they are taps, clean strikes, heavy strikes, followed by submission attempts, control time, etc. The project is easily shrunk if some of the features are not that realistic to implement in 8 months but also easily expanded if we are working faster than expected.

Impressive Documentation Process: Not sure about this one, gave it a 60 since I don't imagine documentation is ever all that interesting, but it is pulled up by the fact that it concerns 2 subjects a lot of people most likely will find interesting (MMA and machine learning). We will go over the thought process of categorizing landed strikes, mma techniques and such so I think it wouldn't be terribly boring.

Interesting to Develop: I only gave this category a 60 partly because I'm not super familiar with how supervised learning would work for video data, but similarly for reasons already listed above, the amount of data to work with could easily be overwhelming hence points were deducted.

Interesting Infrastructure: Deploying the project is easy, just feed it live feed from an MMA fight. -10 points because of lagging issues that occur when you watch on a russian link.

Impressive Infrastructure: gave this a 90 because scoring an old fight is still impressive, but scoring an on going fight would be truly a monumental achievement.

Supervisor(s)/Stakeholder(s) Availability: 35 because while we could have an actual MMA fighter for talks, it would be hard to actually talk to the UFC or a big MMA organization, or a sanctioning body/sports commission. The sports commissions would probably be pissed off at the project.